requirements.txt
flask
boltiotai
python-dotenv
flask_cors
gunicorn
pymongo